backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
gunicorn==23.0.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-debug-toolbar==6.2.0
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.11.0
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
requests==2.31.0